feat: alterando tipo de long para int em TemplateBiometriaUsuario (+ interfaces do repositório de exclusão)

+++ src/Backend/ControleDeAcesso26.Domain/Entities/TemplateBiometriaUsuario.cs
@@ -5,8 +5,8 @@ namespace ControleDeAcesso26.Domain.Entities
 {
     public class TemplateBiometriaUsuario : EntityBase
     {
-        public required long IdSensor1 { get; set; }
-        public long? IdSensor2 { get; set; }
+        public required int IdSensor1 { get; set; }
+        public int? IdSensor2 { get; set; }
         public required byte[] Template { get; set; }
         public required long IdUsuario { get; set; }
